requirements.txt
flask
requests
beautifulsoup4
scikit-learn
pandas
python-whois
python_dateutil
gunicorn